+++ frontend-teesa/src/Components/Login/Login.jsx
@@ -137,7 +137,7 @@ const Login = () => {
     }
   }, [isUserLoaded, navigate]);
 
-  //Prueba Token
+  //*Prueba Token
 
   // const fetchData = async () => {
   //   try {
@@ -150,10 +150,10 @@ const Login = () => {
   //       }
   //     );
   //     console.log(response.data);
-  //      Aquí puedes manejar la respuesta de la API
+  //       Aquí puedes manejar la respuesta de la API
   //   } catch (error) {
   //     console.error(error.response.data);
-  //      Aquí puedes manejar cualquier error ocurrido durante la solicitud
+  //       Aquí puedes manejar cualquier error ocurrido durante la solicitud
   //   }
   // };
