+++ frontend-teesa/src/views/Checkout/CheckoutSucess.jsx
@@ -14,23 +14,15 @@ const CheckoutSucess = () => {
     navigate('/home');
   };
 
-  //Tareas:
+  //Enviar Mail:
 
-  //TODO: 0. Preeliminar: Verificar si funciona mi alternativa para guardar email en Query.
-  //*1.Verifico si está logeado (Google o Nuestro)
-  //*2. Si sí está logeado, saco el nombre y el email (Google o nuestro).
-  //*3. Con el nombre y el user, le envío el mail.
-
-  //Traer Data del User - Nuestro Login y Register (si no funciona así lo traemos de cookies).
   const allUser = useSelector((state) => state.userState);
-  //1. - Si no detecta el user de Google - comprobamos con las cookies.
+
   useEffect(() => {
     if (allUser.user) {
-      //2.
       const user_name = allUser.userData.userName;
       const user_email = allUser.userData.userEmail;
       console.log(user_email);
-      //3.
       emailjs
         .send(
           'service_2rp9duo',
